pytest(fix[doctest-filter]): Apply control doctest filter in plugin
why: Ensure doctest filtering works for downstream users via the pytest plugin entry point.
what:
- Move control-engine doctest filtering hook into libtmux.pytest_plugin
- Add pytester coverage that doctests are filtered under --engine=control
- Drop redundant hook from repo conftest
