+++ src/libtmux/pytest_plugin.py
@@ -451,3 +451,26 @@ def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
         else:
             params = [metafunc.config.getoption("--engine")]
         metafunc.parametrize("engine_name", params, indirect=True)
+
+
+def pytest_collection_modifyitems(
+    config: pytest.Config,
+    items: list[pytest.Item],
+) -> None:
+    """Filter out doctests when running with control engine.
+
+    Remove doctests from collection to avoid pytest's _use_item_location
+    bug: DoctestItem.reportinfo() returns None lineno for fixture doctests,
+    which triggers assertion failure in _pytest/reports.py:420 when skipped.
+    """
+    engine_opt = config.getoption("--engine", default="subprocess")
+    if engine_opt != "control":
+        return
+
+    try:
+        from _pytest.doctest import DoctestItem
+    except Exception:  # pragma: no cover - defensive
+        return
+
+    # Filter out DoctestItems - can't use skip markers due to pytest bug
+    items[:] = [item for item in items if not isinstance(item, DoctestItem)]
